feat: implement thresholds banner functionality and remove notifications banner (PIN-9973)
