+++ src/api/notification/notification.services.ts
@@ -1,11 +1,6 @@
 import axiosInstance from '@/config/axios'
-import axios from 'axios'
-import {
-  APP_MODE,
-  BACKEND_FOR_FRONTEND_URL,
-  FEATURE_FLAG_NOTIFICATION_CONFIG,
-  INTEROP_RESOURCES_BASE_URL,
-} from '@/config/env'
+
+import { BACKEND_FOR_FRONTEND_URL, FEATURE_FLAG_NOTIFICATION_CONFIG } from '@/config/env'
 import type {
   GetNotificationsParams,
   Notifications,
@@ -16,7 +11,6 @@ import type {
   NotificationsCountBySection,
   MarkNotificationsAsReadPayload,
 } from '../api.generatedTypes'
-import type { BannerData } from '@/hooks/bannerHooks/utils'
 
 async function getUserNotificationsList(params: GetNotificationsParams) {
   const response = await axiosInstance.get<Notifications>(
@@ -66,27 +60,6 @@ async function deleteNotifications(payload: { ids: string[] }) {
   return await axiosInstance.delete<void>(`${BACKEND_FOR_FRONTEND_URL}/inAppNotifications`, {
     data: payload,
   })
-}
-
-async function getNotificationsBannerConfigJson() {
-  // Mock data in development to avoid CORS issues with S3
-  if (APP_MODE === 'development') {
-    return {
-      start: {
-        date: '2023-05-25',
-        time: '10:47',
-      },
-      end: {
-        date: '2028-05-25',
-        time: '12:47',
-      },
-    } as BannerData
-  }
-
-  const response = await axios.get<BannerData>(
-    `${INTEROP_RESOURCES_BASE_URL}/notifications-window/data.json`
-  )
-  return response.data
 }
 
 async function markNotificationsAsReadByEntityId({ entityId }: { entityId: string }) {
@@ -179,7 +152,6 @@ export const NotificationServices = {
   markBulkAsNotRead,
   deleteNotification,
   deleteNotifications,
-  getNotificationsBannerConfigJson,
   markNotificationsAsReadByEntityId,
   getInAppNotificationsCount,
 }
