+++ src/static/locales/en/shared-components.json
@@ -332,11 +332,9 @@
     "bodySingleDay": "PDND Interoperabilità will undergo scheduled maintenance on {{maintenanceStartDay}} at {{maintenanceStartHour}} for approximately {{hoursDuration}} hours. We apologize for any inconvenience caused.",
     "bodyMultipleDay": "PDND Interoperabilità will undergo scheduled maintenance from {{maintenanceStartHour}} on {{maintenanceStartDay}} to {{maintenanceEndHour}} on {{maintenanceEndDay}}. We apologize for any inconvenience caused."
   },
-  "notificationsBanner": {
-    "title": "Notifications are here!",
-    "body": "We will inform you when there are news of your interest directly on the platform or via personal email. Check the manual and set your preferences now.",
-    "action1Label": "Configure notifications",
-    "action2Label": "Check the manual"
+  "thresholdsBanner": {
+    "title": "News on PDND: personalized thresholds",
+    "body": "Now providers can customize the number of API calls based on the certified attribute of the consumer. With the release of this new feature, the e-service creation flow has been updated."
   },
   "attributeContainer": {
     "attributeIdLabel": "Attribute ID (attributeId)",
