+++ src/static/locales/it/shared-components.json
@@ -332,11 +332,9 @@
     "bodySingleDay": "Interoperabilità PDND sarà sottoposto a manutenzione programmata il {{maintenanceStartDay}} alle {{maintenanceStartHour}} per circa {{hoursDuration}} ore. Ci scusiamo per gli eventuali disservizi causati.",
     "bodyMultipleDay": "Interoperabilità PDND sarà sottoposto a manutenzione programmata dalle {{maintenanceStartHour}} del {{maintenanceStartDay}} alle {{maintenanceEndHour}} del {{maintenanceEndDay}}. Ci scusiamo per gli eventuali disservizi causati."
   },
-  "notificationsBanner": {
-    "title": "Sono arrivate le notifiche!",
-    "body": "Ti informeremo quando ci sono novità di tuo interesse direttamente in piattaforma o tramite e-mail personale. Consulta il manuale e configura subito le tue preferenze.",
-    "action1Label": "Configura le notifiche",
-    "action2Label": "Consulta il manuale"
+  "thresholdsBanner": {
+    "title": "Novità su PDND: soglie personalizzate",
+    "body": "Adesso gli erogatori possono personalizzare il numero di chiamate API in base all’attributo certificato del fruitore. Con il rilascio di questa nuova funzionalità, il flusso di creazione e-service è stato aggiornato."
   },
   "attributeContainer": {
     "attributeIdLabel": "ID dell'attributo (attributeId)",
